v0.1.2
Changes:
- Renamed gui.py to main.py for clarity.
- Removed capitalization logic. See notes.
- Added simple menu bar with File -> Exit.
- Fleshed out TODO list. Very important.

Notes:
The program was constantly juggling between lowercase category keys and their uppercase counterparts displayed on the GUI. This made no sense because the goal is to have the user be able to save their own transaction categories (misspellings and all), which means the key strings should always be the same wherever they are: in the GUI, in the CSV file, and in the saved category dictionary.

So I removed all capitalization logic and made sure all keys in categories.py appeared the way I wanted them to (previously were all lowercase). This has the added bonus of making things a lot simpler and avoiding potential name errors whenever writing new lines of code that handle the dict.

It also required some tedious renaming, so I created a script for capitalizing dictionary keys. You can see it over at https://github.com/Clovis5119/small_scripts